Good job on this tiny lib! I tried my best, but didn't manage to find what can be improved.

Also I checked a promising alternative to `fastjson`, i.e `jsoniter` which cockroachdb found to be [the best for their case](https://www.cockroachlabs.com/blog/high-performance-json-parsing/). But its `Iterator` doesn't support looking forward, so you don't know how long an array or an object is, which is crucial for messagepack.

Benchmark on main(old) vs this PR(new). The performance virtually didn't change:

```
goos: darwin
goarch: arm64
pkg: github.com/makasim/jsontomsgp
                  │     old     │                new                 │
                  │   sec/op    │   sec/op     vs base               │
CopyJSONToMsgp-10   421.8n ± 1%   415.1n ± 2%  -1.60% (p=0.000 n=30)
```