@makasim 2023-12-06
One possible improvement could be use of for loop instead of recursion.